getting some of the functionality done
